requirements.txt
altgraph==0.16.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
boto3==1.9.210
botocore==1.12.210
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
docutils==0.14
entrypoints==0.3
fbs==0.8.3
future==0.17.1
gevent==1.4.0
greenlet==0.4.15
grequests==0.4.0
idna==2.8
imageio==2.5.0
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.14.1
Jinja2==2.10.1
jmespath==0.9.3
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
macholib==1.11
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mpmath==1.1.0
nbconvert==5.6.0
nbformat==4.4.0
nose==1.3.7
notebook==6.0.0
numpy==1.17.0
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
pefile==2018.8.8
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
PyInstaller==3.5
pyparsing==2.4.2
PyQt5==5.9.2
PyQt5-sip==4.19.18
pyrsistent==0.15.4
python-dateutil==2.7.5
pytz==2019.2
pyzmq==18.1.0
qtconsole==4.5.2
requests==2.22.0
s3transfer==0.2.1
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
sympy==1.4
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
